# demo-app/docs/API_SPEC.yaml
openapi: "3.1.0"
info:
  title: Demo App API
  version: 1.0.0
  description: |
    REST API for the Angular 21 Enterprise Demo App.

    ## Authentication
    All protected endpoints use an HttpOnly JWT cookie (`access-token`, RS256, 15-min TTL)
    set by `POST /auth/login` and rotated by `POST /auth/refresh`.
    The browser sends it automatically when Angular `HttpClient` is configured with
    `withCredentials: true`. No `Authorization` header is used.

    ## CSRF
    CSRF tokens are **disabled**. Protection is provided by `SameSite=Strict` on both
    auth cookies combined with the strict CORS origin allowlist.

    ## Pagination
    List endpoints accept `page` (0-based) and `pageSize` query parameters.
    The total item count is returned in the `X-Total-Count` response header.
    Response bodies for list endpoints are plain JSON arrays.

    ## Errors
    All error responses share the same shape:
    ```json
    { "code": "ERROR_CODE", "message": "human-readable description",
      "fields": [{ "field": "email", "message": "must be a valid email" }] }
    ```
servers:
  - url: https://api.company.com/api/v1
    description: Production
  - url: https://staging.company.com/api/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - cookieAuth: []
...
paths:
...
  /auth/password:
    post:
      tags: [Auth]
      summary: Change the authenticated user's password
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: Validation error or wrong current password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                wrongPassword:
                  value:
                    code: WRONG_PASSWORD
                    message: Current password is incorrect
                tooShort:
                  value:
                    code: VALIDATION_ERROR
                    message: Validation failed
                    fields:
                      - field: newPassword
                        message: must be at least 6 characters

  # ── Users ─────────────────────────────────────────────────────────────────
  /users:
    get:
      tags: [Users]
      summary: List users
      operationId: listUsers
      x-permission: usersView
      parameters:
        - name: q
          in: query
          description: Search term matched against name and email
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paginated user list
          headers:
            X-Total-Count:
              $ref: '#/components/headers/XTotalCount'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Users]
      summary: Create a user
      operationId: createUser
      x-permission: usersCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: CONFLICT
                message: A user with this email already exists
        '422':
          $ref: '#/components/responses/ValidationError'
  /users/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'

    patch:
      tags: [Users]
      summary: Update a user
      operationId: updateUser
      x-permission: usersEdit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Users]
      summary: Delete a user
      operationId: deleteUser
      x-permission: usersDelete
      responses:
        '204':
          description: User deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Cannot delete your own account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: CONFLICT
                message: You cannot delete your own account

  # ── Roles ─────────────────────────────────────────────────────────────────
...
  /reports/upload-trend:
    get:
      tags: [Reports]
      summary: Monthly upload trend for the last 12 months
      operationId: getUploadTrend
      x-permission: rolesView
      description: Returns one series per category. Results are cached in Redis (TTL 5 min).
      responses:
        '200':
          description: Upload trend data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadTrendResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /reports/role-distribution:
    get:
      tags: [Reports]
      summary: User count per role
      operationId: getRoleDistribution
      x-permission: rolesView
      responses:
        '200':
          description: Role distribution data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDistributionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /profile:
    get:
      tags: [Profile]
      summary: Get the authenticated user's profile
      operationId: getProfile
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      tags: [Profile]
      summary: Update display name
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdateRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'

  # ── Audit ─────────────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access-token
      description: |
        HttpOnly JWT cookie (RS256, 15-min TTL). Set by `POST /auth/login` and
        rotated by `POST /auth/refresh`. Angular `HttpClient` sends it automatically
        when configured with `withCredentials: true`. Never accessible to JavaScript.

  # ── Headers ───────────────────────────────────────────────────────────────
  headers:
    XTotalCount:
      description: Total number of items matching the query (before pagination)
      schema:
        type: integer
        example: 42

  # ── Parameters ────────────────────────────────────────────────────────────
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Resource UUID
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Zero-based page number
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        enum: [5, 10, 25, 50]
        default: 10
      description: Number of items per page
    SortParam:
      name: sort
      in: query
      schema:
        type: string
        example: createdAt,desc
      description: "Sort field and direction separated by comma. Example: `createdAt,desc`"

  # ── Reusable Responses ────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid authentication cookie
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: Authentication required
    Forbidden:
      description: Authenticated but insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: FORBIDDEN
            message: You do not have permission to perform this action
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: NOT_FOUND
            message: The requested resource was not found
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: VALIDATION_ERROR
            message: Validation failed
            fields:
              - field: email
                message: must be a valid email address
...
  schemas:
...
    UserStatus:
      type: string
      enum: [active, inactive, suspended]
      description: Lifecycle status of a user account
...
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - BUILT_IN_ROLE
        - WRONG_PASSWORD
        - UPLOAD_EXPIRED
        - RATE_LIMITED
        - INTERNAL_ERROR

    # ── Core Entities ────────────────────────────────────────────────────────
    RolePermissions:
      type: object
      required:
        - usersView
        - usersCreate
        - usersEdit
        - usersDelete
        - rolesView
        - rolesEdit
      properties:
        usersView:
          type: boolean
        usersCreate:
          type: boolean
        usersEdit:
          type: boolean
        usersDelete:
          type: boolean
        rolesView:
          type: boolean
        rolesEdit:
          type: boolean
    User:
      type: object
      required: [id, email, displayName, status, roleId, roleName, permissions, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        email:
          type: string
          format: email
        displayName:
          type: string
        status:
          $ref: '#/components/schemas/UserStatus'
        roleId:
          type: string
          format: uuid
        roleName:
          type: string
        permissions:
          $ref: '#/components/schemas/RolePermissions'
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    UserCreateRequest:
      type: object
      required: [email, displayName, password, roleId]
      properties:
        email:
          type: string
          format: email
        displayName:
          type: string
          minLength: 2
          maxLength: 100
        password:
          type: string
          minLength: 6
          format: password
        roleId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/UserStatus'
    UserUpdateRequest:
      type: object
      properties:
        displayName:
          type: string
          minLength: 2
          maxLength: 100
        roleId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/UserStatus'
...
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          minLength: 6
          format: password
    ProfileUpdateRequest:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
          minLength: 2
          maxLength: 100

    # ── Report Responses ─────────────────────────────────────────────────────
    UploadTrendResponse:
      type: object
      required: [categories, series]
      description: Data for the upload trend area-spline chart (last 12 months).
      properties:
        categories:
          type: array
          items:
            type: string
          example: ["Jun 2024","Jul 2024","Aug 2024","Sep 2024","Oct 2024","Nov 2024","Dec 2024","Jan 2025","Feb 2025","Mar 2025","Apr 2025","May 2025"]
          description: Month labels. Last entry is always the current month.
        series:
          type: array
          items:
            type: object
            required: [name, data]
            properties:
              name:
                type: string
                description: Category name
              data:
                type: array
                items:
                  type: integer
                  minimum: 0
                description: Upload count per month (12 values, aligned with categories)
...
    RoleDistributionResponse:
      type: object
      required: [data]
      description: Data for the role distribution donut chart.
      properties:
        data:
          type: array
          items:
            type: object
            required: [name, value]
            properties:
              name:
                type: string
                description: Role name
              value:
                type: integer
                minimum: 0
                description: Number of users with this role

    # ── Audit ────────────────────────────────────────────────────────────────
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Human-readable explanation
        fields:
          type: array
          description: Per-field validation errors (present only for VALIDATION_ERROR)
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
                description: JSON path to the invalid field
              message:
                type: string
                description: Why the field failed validation

    # ── Health ───────────────────────────────────────────────────────────────
...